Refactored colormap to an utin in style
