+++ packages/gdl-frontend/components/Level/LevelHR.js
@@ -10,15 +10,7 @@ import * as React from 'react';
 import styled from 'react-emotion';
 
 import type { ReadingLevel } from '../../types';
-
-const colorMap = {
-  '1': '#5DD0C1',
-  '2': '#84CB65',
-  '3': '#F1C528',
-  '4': '#FA9F28',
-  'read-aloud': '#F56324',
-  decodable: '#AB86CD'
-};
+import colorMap from '../../style/colorMapping';
 
 const Hr = styled('hr')`
   border: 2px solid ${p => (p.level ? colorMap[p.level] : '#D2D2D2')};
