+++ packages/gdl-frontend/style/colorMapping.js
@@ -1,0 +1,18 @@
+// @flow
+/**
+ * Part of GDL gdl-frontend.
+ * Copyright (C) 2017 GDL
+ *
+ * See LICENSE
+ */
+
+const colorMap = {
+  '1': '#5DD0C1',
+  '2': '#84CB65',
+  '3': '#F1C528',
+  '4': '#FA9F28',
+  'read-aloud': '#F56324',
+  decodable: '#AB86CD'
+};
+
+export default colorMap;
